Revise "Update URI" descriptor in plugin basics
The description of the "Update URI" field in the plugin basics section has been clarified. The update specifically stresses avoiding usage of the update URI for plugins hosted in the WordPress.org Plugin Directory. This change aims to prevent any accidental overwrites of these plugins due to confusion with similarly named plugins.