Unnecessary line breaks appear to be inserted at the end of lines.
At the end of line 76, an unnecessary line break appears to have been inserted.
The content of line 77 should continue at the end of line 76.
line 76-77: ‘When registering you provide a friendly name for the eraser (to aid in debugging – this friendly name is not shown to anyone at this time)
and the callback, e.g.’